Fix a small bug for positional argument default value being None
For `F.scaled_dot_product_attention` we have a positional argument:

`Tensor? attn_mask=None`

This is not handled correctly in `python_variables.cpp`, and it is
causing issue in tracing.

Fix it so that `is_default` lambda  correctly returns true when the default value is None

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5a280b4d902f76c20b2f8761cb6c559b108681e4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113102